Module gsl_rust::distribution

Functions

bernoulli_pdf

This function computes the probability $p(k)$ of obtaining k from a Bernoulli distribution with probability parameter p.

beta_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the beta distribution with parameters a and b.

beta_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the beta distribution with parameters a and b.

beta_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the beta distribution with parameters a and b.

beta_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the beta distribution with parameters a and b.

beta_pdf

This function computes the probability density $p(x)$ at x for a beta distribution with parameters a and b.

binomial_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the binomial distribution with parameters p and n.

binomial_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the binomial distribution with parameters p and n.

binomial_pdf

This function computes the probability $p(k)$ of obtaining k from a binomial distribution with parameters p and n.

bivariate_gaussian_pdf

This function computes the probability density $p(x,y)$ at (x, y) for a bivariate Gaussian distribution with standard deviations sigma_x, sigma_y and correlation coefficient rho.

cauchy_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Cauchy distribution with scale parameter a.

cauchy_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Cauchy distribution with scale parameter a.

cauchy_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Cauchy distribution with scale parameter a.

cauchy_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Cauchy distribution with scale parameter a.

cauchy_pdf

This function computes the probability density $p(x)$ at x for a Cauchy distribution with scale parameter a.

chi2

This function returns a random variate from the chi-squared distribution with nu degrees of freedom.

chi2_pdf

This function computes the probability density $p(x)$ at x for a chi-squared distribution with nu degrees of freedom.

chi_squared_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the chi-squared distribution with nu degrees of freedom.

chi_squared_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the chi-squared distribution with nu degrees of freedom.

chi_squared_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the chi-squared distribution with nu degrees of freedom.

chi_squared_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the chi-squared distribution with nu degrees of freedom.

draw_sample_from_bernoulli

This function returns either 0 or 1, the result of a Bernoulli trial with probability p.

draw_sample_from_beta

This function returns a random variate from the beta distribution.

draw_sample_from_binomial

This function returns a random integer from the binomial distribution, the number of successes in n independent trials with probability p.

draw_sample_from_cauchy

This function returns a random variate from the Cauchy distribution with scale parameter a.

draw_sample_from_exponential

This function returns a random variate from the exponential distribution with mean mu.

draw_sample_from_exppow

This function returns a random variate from the exponential power distribution with scale parameter a and exponent b.

draw_sample_from_flat

This function returns a random variate from the flat (uniform) distribution from a to b.

draw_sample_from_gamma

This function returns a random variate from the gamma distribution.

draw_sample_from_gamma_knuth

This function returns a gamma variate using the algorithms from Knuth (vol 2).

draw_sample_from_gaussian

This function returns a Gaussian random variate, with mean zero and standard deviation sigma.

draw_sample_from_gaussian_ratio_method

This function computes a Gaussian random variate using the alternative Marsaglia-Tsang ziggurat and Kinderman-Monahan-Leva ratio methods.

draw_sample_from_gaussian_tail

This function provides random variates from the upper tail of a Gaussian distribution with standard deviation sigma.

draw_sample_from_gaussian_ziggurat

This function computes a Gaussian random variate using the alternative Marsaglia-Tsang ziggurat and Kinderman-Monahan-Leva ratio methods.

draw_sample_from_geometric

This function returns a random integer from the geometric distribution, the number of independent trials with probability p until the first success.

draw_sample_from_gumbel1

This function returns a random variate from the Type-1 Gumbel distribution.

draw_sample_from_gumbel2

This function returns a random variate from the Type-2 Gumbel distribution.

draw_sample_from_hypergeometric

This function returns a random integer from the hypergeometric distribution.

draw_sample_from_landau

This function returns a random variate from the Landau distribution.

draw_sample_from_laplace

This function returns a random variate from the Laplace distribution with width a.

draw_sample_from_levy

This function returns a random variate from the Levy symmetric stable distribution with scale c and exponent alpha.

draw_sample_from_levy_skew

This function returns a random variate from the Levy skew stable distribution with scale c, exponent alpha and skewness parameter beta.

draw_sample_from_logarithmic

This function returns a random integer from the logarithmic distribution.

draw_sample_from_logistic

This function returns a random variate from the logistic distribution.

draw_sample_from_lognormal

This function returns a random variate from the lognormal distribution.

draw_sample_from_negative_binomial

This function returns a random integer from the negative binomial distribution, the number of failures occurring before n successes in independent trials with probability p of success.

draw_sample_from_pareto

This function returns a random variate from the Pareto distribution of order a.

draw_sample_from_pascal

This function returns a random integer from the Pascal distribution.

draw_sample_from_poisson

This function returns a random integer from the Poisson distribution with mean mu.

draw_sample_from_rayleigh

This function returns a random variate from the Rayleigh distribution with scale parameter sigma.

draw_sample_from_rayleigh_tail

This function returns a random variate from the tail of the Rayleigh distribution with scale parameter sigma and a lower limit of a.

draw_sample_from_ugaussian

These functions compute results for the unit Gaussian distribution.

draw_sample_from_ugaussian_ratio_method

These functions compute results for the unit Gaussian distribution.

draw_sample_from_ugaussian_tail

These functions compute results for the tail of a unit Gaussian distribution.

draw_sample_from_weibull

This function returns a random variate from the Weibull distribution.

draw_samples_from_bernoulli

This function returns either 0 or 1, the result of nr_of_samples Bernoulli trials with probability p.

draw_samples_from_beta

This function returns nr_of_samples random variates from the beta distribution.

draw_samples_from_binomial

This function returns nr_of_samples random integers from the binomial distribution, the number of successes in n independent trials with probability p.

draw_samples_from_cauchy

This function returns nr_of_samples random variates from the Cauchy distribution with scale parameter a.

draw_samples_from_exponential

This function returns nr_of_samples random variates from the exponential distribution with mean mu.

draw_samples_from_exppow

This function returns nr_of_samples random variates from the exponential power distribution with scale parameter a and exponent b.

draw_samples_from_flat

This function returns nr_of_samples random variates from the flat (uniform) distribution from a to b.

draw_samples_from_gamma

This function returns nr_of_samples random variates from the gamma distribution.

draw_samples_from_gamma_knuth

This function returns a gamma variate using the algorithms from Knuth (vol 2).

draw_samples_from_gaussian

This function returns returns nr_of_samples Gaussian random variates, with mean zero and standard deviation sigma.

draw_samples_from_gaussian_ratio_method

This function computes returns nr_of_samples Gaussian random variates using the alternative Marsaglia-Tsang ziggurat and Kinderman-Monahan-Leva ratio methods.

draw_samples_from_gaussian_tail

This function provides random variates from the upper tail of a Gaussian distribution with standard deviation sigma.

draw_samples_from_gaussian_ziggurat

This function computes returns nr_of_samples Gaussian random variates using the alternative Marsaglia-Tsang ziggurat and Kinderman-Monahan-Leva ratio methods.

draw_samples_from_geometric

This function returns nr_of_samples random integers from the geometric distribution, the number of independent trials with probability p until the first success.

draw_samples_from_gumbel1

This function returns nr_of_samples random variates from the Type-1 Gumbel distribution.

draw_samples_from_gumbel2

This function returns nr_of_samples random variates from the Type-2 Gumbel distribution.

draw_samples_from_hypergeometric

This function returns nr_of_samples random integers from the hypergeometric distribution.

draw_samples_from_landau

This function returns nr_of_samples random variates from the Landau distribution.

draw_samples_from_laplace

This function returns nr_of_samples random variates from the Laplace distribution with width a.

draw_samples_from_levy

This function returns nr_of_samples random variates from the Levy symmetric stable distribution with scale c and exponent alpha.

draw_samples_from_levy_skew

This function returns nr_of_samples random variates from the Levy skew stable distribution with scale c, exponent alpha and skewness parameter beta.

draw_samples_from_logarithmic

This function returns nr_of_samples random integers from the logarithmic distribution.

draw_samples_from_logistic

This function returns nr_of_samples random variates from the logistic distribution.

draw_samples_from_lognormal

This function returns nr_of_samples random variates from the lognormal distribution.

draw_samples_from_negative_binomial

This function returns nr_of_samples random integers from the negative binomial distribution, the number of failures occurring before n successes in independent trials with probability p of success.

draw_samples_from_pareto

This function returns nr_of_samples random variates from the Pareto distribution of order a.

draw_samples_from_pascal

This function returns nr_of_samples random integers from the Pascal distribution.

draw_samples_from_poisson

This function returns nr_of_samples random integers from the Poisson distribution with mean mu.

draw_samples_from_rayleigh

This function returns nr_of_samples random variates from the Rayleigh distribution with scale parameter sigma.

draw_samples_from_rayleigh_tail

This function returns nr_of_samples random variates from the tail of the Rayleigh distribution with scale parameter sigma and a lower limit of a.

draw_samples_from_ugaussian

These functions compute results for the unit Gaussian distribution.

draw_samples_from_ugaussian_ratio_method

These functions compute results for the unit Gaussian distribution.

draw_samples_from_ugaussian_tail

These functions compute results for the tail of a unit Gaussian distribution.

draw_samples_from_weibull

This function returns nr_of_samples random variates from the Weibull distribution.

exponential_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the exponential distribution with mean mu.

exponential_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the exponential distribution with mean mu.

exponential_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the exponential distribution with mean mu.

exponential_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the exponential distribution with mean mu.

exponential_pdf

This function computes the probability density $p(x)$ at x for an exponential distribution with mean mu.

exppow_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ for the exponential power distribution with parameters a and b.

exppow_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ for the exponential power distribution with parameters a and b.

exppow_pdf

This function computes the probability density $p(x)$ at x for an exponential power distribution with scale parameter a and exponent b, using the formula given above.

f_dist

This function returns a random variate from the F-distribution with degrees of freedom nu1 and nu2.

f_dist_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the F-distribution with nu1 and nu2 degrees of freedom.

f_dist_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the F-distribution with nu1 and nu2 degrees of freedom.

f_dist_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the F-distribution with nu1 and nu2 degrees of freedom.

f_dist_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the F-distribution with nu1 and nu2 degrees of freedom.

f_dist_pdf

This function computes the probability density $p(x)$ at x for an F-distribution with nu1 and nu2 degrees of freedom, using the formula given above.

flat_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for a uniform distribution from a to b.

flat_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for a uniform distribution from a to b.

flat_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for a uniform distribution from a to b.

flat_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for a uniform distribution from a to b.

flat_pdf

This function computes the probability density $p(x)$ at x for a uniform distribution from a to b.

gamma_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the gamma distribution with parameters a and b.

gamma_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the gamma distribution with parameters a and b.

gamma_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the gamma distribution with parameters a and b.

gamma_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the gamma distribution with parameters a and b.

gamma_pdf

This function computes the probability density $p(x)$ at x for a gamma distribution with parameters a and b.

gaussian_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_pdf

This function computes the probability density $p(x)$ at x for a Gaussian distribution with standard deviation sigma.

gaussian_tail_pdf

This function computes the probability density $p(x)$ at x for a Gaussian tail distribution with standard deviation sigma and lower limit a, using the formula given above.

geometric_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the geometric distribution with parameter p.

geometric_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the geometric distribution with parameter p.

geometric_pdf

This function computes the probability $p(k)$ of obtaining k from a geometric distribution with probability parameter p.

gumbel1_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-1 Gumbel distribution with parameters a and b.

gumbel1_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-1 Gumbel distribution with parameters a and b.

gumbel1_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-1 Gumbel distribution with parameters a and b.

gumbel1_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-1 Gumbel distribution with parameters a and b.

gumbel1_pdf

This function computes the probability density $p(x)$ at x for a Type-1 Gumbel distribution with parameters a and b.

gumbel2_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_pdf

This function computes the probability density $p(x)$ at x for a Type-2 Gumbel distribution with parameters a and b.

hypergeometric_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the hypergeometric distribution with parameters n1, n2 and t.

hypergeometric_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the hypergeometric distribution with parameters n1, n2 and t.

hypergeometric_pdf

This function computes the probability $p(k)$ of obtaining k from a hypergeometric distribution with parameters n1, n2, t, using the formula given above.

landau_pdf

This function computes the probability density $p(x)$ at x for the Landau distribution using an approximation to the formula given above.

laplace_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Laplace distribution with width a.

laplace_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Laplace distribution with width a.

laplace_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Laplace distribution with width a.

laplace_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Laplace distribution with width a.

laplace_pdf

This function computes the probability density $p(x)$ at x for a Laplace distribution with width a.

logarithmic_pdf

This function computes the probability $p(k)$ of obtaining k from a logarithmic distribution with probability parameter p, using the formula given above.

logistic_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the logistic distribution with scale parameter a.

logistic_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the logistic distribution with scale parameter a.

logistic_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the logistic distribution with scale parameter a.

logistic_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the logistic distribution with scale parameter a.

logistic_pdf

This function computes the probability density $p(x)$ at x for a logistic distribution with scale parameter a.

lognormal_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_pdf

This function computes the probability density $p(x)$ at x for a lognormal distribution with parameters zeta and sigma.

negative_binomial_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the negative binomial distribution with parameters p and n.

negative_binomial_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the negative binomial distribution with parameters p and n.

negative_binomial_pdf

This function computes the probability $p(k)$ of obtaining k from a negative binomial distribution with parameters p and n, using the formula given above.

pareto_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Pareto distribution with exponent a and scale b.

pareto_pdf

This function computes the probability density $p(x)$ at x for a Pareto distribution with exponent a and scale b.

pascal_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the Pascal distribution with parameters p and n.

pascal_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the Pascal distribution with parameters p and n.

pascal_pdf

This function computes the probability $p(k)$ of obtaining k from a Pascal distribution with parameters p and n.

poisson_P_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the Poisson distribution with parameter mu.

poisson_Q_cdf

These functions compute the cumulative distribution functions $P(k)$, $Q(k)$ for the Poisson distribution with parameter mu.

poisson_pdf

This function computes the probability $p(k)$ of obtaining k from a Poisson distribution with mean mu.

rayleigh_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_pdf

This function computes the probability density $p(x)$ at x for a Rayleigh distribution with scale parameter sigma.

rayleigh_tail_pdf

This function computes the probability density $p(x)$ at x for a Rayleigh tail distribution with scale parameter sigma and lower limit a, using the formula given above.

student_t

This function returns a random variate from the t-distribution.

student_t_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the t-distribution with nu degrees of freedom.

student_t_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the t-distribution with nu degrees of freedom.

student_t_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the t-distribution with nu degrees of freedom.

student_t_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the t-distribution with nu degrees of freedom.

student_t_pdf

This function computes the probability density $p(x)$ at x for a t-distribution with nu degrees of freedom.

ugaussian_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the unit Gaussian distribution.

ugaussian_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the unit Gaussian distribution.

ugaussian_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the unit Gaussian distribution.

ugaussian_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the unit Gaussian distribution.

ugaussian_pdf

These functions compute results for the unit Gaussian distribution.

ugaussian_tail_pdf

These functions compute results for the tail of a unit Gaussian distribution.

weibull_P_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Weibull distribution with scale a and exponent b.

weibull_Pinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Weibull distribution with scale a and exponent b.

weibull_Q_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Weibull distribution with scale a and exponent b.

weibull_Qinv_cdf

These functions compute the cumulative distribution functions $P(x)$, $Q(x)$ and their inverses for the Weibull distribution with scale a and exponent b.

weibull_pdf

This function computes the probability density $p(x)$ at x for a Weibull distribution with scale a and exponent b.