Module gsl_rust::special

Modules

special_function_test

Functions

Chi

These routines compute the integral

Chi_e

These routines compute the integral

Ci

These routines compute the Cosine integral

Ci_e

These routines compute the Cosine integral

Shi

These routines compute the integral

Shi_e

These routines compute the integral

Si

These routines compute the Sine integral

Si_e

These routines compute the Sine integral

angle_restrict_pos

These routines force the angle theta to lie in the range $[0, 2\pi)$.

angle_restrict_symm

These routines force the angle theta to lie in the range $(-\pi,\pi]$.

atanint

These routines compute the Arctangent integral, which is defined as

atanint_e

These routines compute the Arctangent integral, which is defined as

bessel_I0

These routines compute the regular modified cylindrical Bessel function of zeroth order, $I_0(x)$.

bessel_I0_e

These routines compute the regular modified cylindrical Bessel function of zeroth order, $I_0(x)$.

bessel_I0_scaled

These routines compute the scaled regular modified cylindrical Bessel function of zeroth order $\exp(-|x|) I_0(x)$.

bessel_I0_scaled_e

These routines compute the scaled regular modified cylindrical Bessel function of zeroth order $\exp(-|x|) I_0(x)$.

bessel_I1

These routines compute the regular modified cylindrical Bessel function of first order, $I_1(x)$.

bessel_I1_e

These routines compute the regular modified cylindrical Bessel function of first order, $I_1(x)$.

bessel_I1_scaled

These routines compute the scaled regular modified cylindrical Bessel function of first order $\exp(-|x|) I_1(x)$.

bessel_I1_scaled_e

These routines compute the scaled regular modified cylindrical Bessel function of first order $\exp(-|x|) I_1(x)$.

bessel_In

These routines compute the regular modified cylindrical Bessel function of order n, $I_n(x)$.

bessel_In_e

These routines compute the regular modified cylindrical Bessel function of order n, $I_n(x)$.

bessel_In_scaled

These routines compute the scaled regular modified cylindrical Bessel function of order n, $\exp(-|x|) I_n(x)$

bessel_In_scaled_e

These routines compute the scaled regular modified cylindrical Bessel function of order n, $\exp(-|x|) I_n(x)$

bessel_Inu

These routines compute the regular modified Bessel function of fractional order $\nu$, $I_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Inu_e

These routines compute the regular modified Bessel function of fractional order $\nu$, $I_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Inu_scaled

These routines compute the scaled regular modified Bessel function of fractional order $\nu$, $\exp(-|x|)I_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Inu_scaled_e

These routines compute the scaled regular modified Bessel function of fractional order $\nu$, $\exp(-|x|)I_\nu(x)$ for $x>0$, $\nu>0$.

bessel_J0

These routines compute the regular cylindrical Bessel function of zeroth order, $J_0(x)$.

bessel_J0_e

These routines compute the regular cylindrical Bessel function of zeroth order, $J_0(x)$.

bessel_J1

These routines compute the regular cylindrical Bessel function of first order, $J_1(x)$.

bessel_J1_e

These routines compute the regular cylindrical Bessel function of first order, $J_1(x)$.

bessel_Jn

These routines compute the regular cylindrical Bessel function of order n, $J_n(x)$.

bessel_Jn_e

These routines compute the regular cylindrical Bessel function of order n, $J_n(x)$.

bessel_Jnu

These routines compute the regular cylindrical Bessel function of fractional order $\nu$, $J_\nu(x)$.

bessel_Jnu_e

These routines compute the regular cylindrical Bessel function of fractional order $\nu$, $J_\nu(x)$.

bessel_K0

These routines compute the irregular modified cylindrical Bessel function of zeroth order, $K_0(x)$, for $x > 0$.

bessel_K0_e

These routines compute the irregular modified cylindrical Bessel function of zeroth order, $K_0(x)$, for $x > 0$.

bessel_K0_scaled

These routines compute the scaled irregular modified cylindrical Bessel function of zeroth order $\exp(x) K_0(x)$ for $x>0$.

bessel_K0_scaled_e

These routines compute the scaled irregular modified cylindrical Bessel function of zeroth order $\exp(x) K_0(x)$ for $x>0$.

bessel_K1

These routines compute the irregular modified cylindrical Bessel function of first order, $K_1(x)$, for $x > 0$.

bessel_K1_e

These routines compute the irregular modified cylindrical Bessel function of first order, $K_1(x)$, for $x > 0$.

bessel_K1_scaled

These routines compute the scaled irregular modified cylindrical Bessel function of first order $\exp(x) K_1(x)$ for $x>0$.

bessel_K1_scaled_e

These routines compute the scaled irregular modified cylindrical Bessel function of first order $\exp(x) K_1(x)$ for $x>0$.

bessel_Kn

These routines compute the irregular modified cylindrical Bessel function of order n, $K_n(x)$, for $x > 0$.

bessel_Kn_e

These routines compute the irregular modified cylindrical Bessel function of order n, $K_n(x)$, for $x > 0$.

bessel_Kn_scaled

These routines compute the scaled irregular modified cylindrical Bessel function of order n, $\exp(x) K_n(x)$, for $x>0$.

bessel_Kn_scaled_e

These routines compute the scaled irregular modified cylindrical Bessel function of order n, $\exp(x) K_n(x)$, for $x>0$.

bessel_Knu

These routines compute the irregular modified Bessel function of fractional order $\nu$, $K_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Knu_e

These routines compute the irregular modified Bessel function of fractional order $\nu$, $K_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Knu_scaled

These routines compute the scaled irregular modified Bessel function of fractional order $\nu$, $\exp(+|x|) K_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Knu_scaled_e

These routines compute the scaled irregular modified Bessel function of fractional order $\nu$, $\exp(+|x|) K_\nu(x)$ for $x>0$, $\nu>0$.

bessel_Y0

These routines compute the irregular cylindrical Bessel function of zeroth order, $Y_0(x)$, for $x>0$.

bessel_Y0_e

These routines compute the irregular cylindrical Bessel function of zeroth order, $Y_0(x)$, for $x>0$.

bessel_Y1

These routines compute the irregular cylindrical Bessel function of first order, $Y_1(x)$, for $x>0$.

bessel_Y1_e

These routines compute the irregular cylindrical Bessel function of first order, $Y_1(x)$, for $x>0$.

bessel_Yn

These routines compute the irregular cylindrical Bessel function of order n, $Y_n(x)$, for $x>0$.

bessel_Yn_e

These routines compute the irregular cylindrical Bessel function of order n, $Y_n(x)$, for $x>0$.

bessel_Ynu

These routines compute the irregular cylindrical Bessel function of fractional order $\nu$, $Y_\nu(x)$.

bessel_Ynu_e

These routines compute the irregular cylindrical Bessel function of fractional order $\nu$, $Y_\nu(x)$.

bessel_i0_scaled

These routines compute the scaled regular modified spherical Bessel function of zeroth order, $\exp(-|x|) i_0(x)$.

bessel_i0_scaled_e

These routines compute the scaled regular modified spherical Bessel function of zeroth order, $\exp(-|x|) i_0(x)$.

bessel_i1_scaled

These routines compute the scaled regular modified spherical Bessel function of first order, $\exp(-|x|) i_1(x)$.

bessel_i1_scaled_e

These routines compute the scaled regular modified spherical Bessel function of first order, $\exp(-|x|) i_1(x)$.

bessel_i2_scaled

These routines compute the scaled regular modified spherical Bessel function of second order, $\exp(-|x|) i_2(x)$

bessel_i2_scaled_e

These routines compute the scaled regular modified spherical Bessel function of second order, $\exp(-|x|) i_2(x)$

bessel_il_scaled

These routines compute the scaled regular modified spherical Bessel function of order l, $\exp(-|x|) i_l(x)$

bessel_il_scaled_e

These routines compute the scaled regular modified spherical Bessel function of order l, $\exp(-|x|) i_l(x)$

bessel_j0

These routines compute the regular spherical Bessel function of zeroth order, $j_0(x) = \sin(x)/x$.

bessel_j0_e

These routines compute the regular spherical Bessel function of zeroth order, $j_0(x) = \sin(x)/x$.

bessel_j1

These routines compute the regular spherical Bessel function of first order, $j_1(x) = (\sin(x)/x - \cos(x))/x$.

bessel_j1_e

These routines compute the regular spherical Bessel function of first order, $j_1(x) = (\sin(x)/x - \cos(x))/x$.

bessel_j2

These routines compute the regular spherical Bessel function of second order, $j_2(x) = ((3/x^2 - 1)\sin(x) - 3\cos(x)/x)/x$.

bessel_j2_e

These routines compute the regular spherical Bessel function of second order, $j_2(x) = ((3/x^2 - 1)\sin(x) - 3\cos(x)/x)/x$.

bessel_jl

These routines compute the regular spherical Bessel function of order l, $j_l(x)$, for $l \geq 0$ and $x \geq 0$.

bessel_jl_e

These routines compute the regular spherical Bessel function of order l, $j_l(x)$, for $l \geq 0$ and $x \geq 0$.

bessel_k0_scaled

These routines compute the scaled irregular modified spherical Bessel function of zeroth order, $\exp(x) k_0(x)$, for $x>0$.

bessel_k0_scaled_e

These routines compute the scaled irregular modified spherical Bessel function of zeroth order, $\exp(x) k_0(x)$, for $x>0$.

bessel_k1_scaled

These routines compute the scaled irregular modified spherical Bessel function of first order, $\exp(x) k_1(x)$, for $x>0$.

bessel_k1_scaled_e

These routines compute the scaled irregular modified spherical Bessel function of first order, $\exp(x) k_1(x)$, for $x>0$.

bessel_k2_scaled

These routines compute the scaled irregular modified spherical Bessel function of second order, $\exp(x) k_2(x)$, for $x>0$.

bessel_k2_scaled_e

These routines compute the scaled irregular modified spherical Bessel function of second order, $\exp(x) k_2(x)$, for $x>0$.

bessel_kl_scaled

These routines compute the scaled irregular modified spherical Bessel function of order l, $\exp(x) k_l(x)$, for $x>0$.

bessel_kl_scaled_e

These routines compute the scaled irregular modified spherical Bessel function of order l, $\exp(x) k_l(x)$, for $x>0$.

bessel_ln_Knu

These routines compute the logarithm of the irregular modified Bessel function of fractional order $\nu$, $\ln(K_\nu(x))$ for $x>0$, $\nu>0$.

bessel_ln_Knu_e

These routines compute the logarithm of the irregular modified Bessel function of fractional order $\nu$, $\ln(K_\nu(x))$ for $x>0$, $\nu>0$.

bessel_y0

These routines compute the irregular spherical Bessel function of zeroth order, $y_0(x) = -\cos(x)/x$.

bessel_y0_e

These routines compute the irregular spherical Bessel function of zeroth order, $y_0(x) = -\cos(x)/x$.

bessel_y1

These routines compute the irregular spherical Bessel function of first order, $y_1(x) = -(\cos(x)/x + \sin(x))/x$.

bessel_y1_e

These routines compute the irregular spherical Bessel function of first order, $y_1(x) = -(\cos(x)/x + \sin(x))/x$.

bessel_y2

These routines compute the irregular spherical Bessel function of second order, $y_2(x) = (-3/x^3 + 1/x)\cos(x) - (3/x^2)\sin(x)$.

bessel_y2_e

These routines compute the irregular spherical Bessel function of second order, $y_2(x) = (-3/x^3 + 1/x)\cos(x) - (3/x^2)\sin(x)$.

bessel_yl

These routines compute the irregular spherical Bessel function of order l, $y_l(x)$, for $l \geq 0$.

bessel_yl_e

These routines compute the irregular spherical Bessel function of order l, $y_l(x)$, for $l \geq 0$.

bessel_zero_J0

These routines compute the location of the s-th positive zero of the Bessel function $J_0(x)$.

bessel_zero_J0_e

These routines compute the location of the s-th positive zero of the Bessel function $J_0(x)$.

bessel_zero_J1

These routines compute the location of the s-th positive zero of the Bessel function $J_1(x)$.

bessel_zero_J1_e

These routines compute the location of the s-th positive zero of the Bessel function $J_1(x)$.

bessel_zero_Jnu

These routines compute the location of the s-th positive zero of the Bessel function $J_\nu(x)$. The current implementation does not support negative values of nu.

bessel_zero_Jnu_e

These routines compute the location of the s-th positive zero of the Bessel function $J_\nu(x)$. The current implementation does not support negative values of nu.

beta

These routines compute the Beta Function, $B(a,b) = \Gamma(a)\Gamma(b)/\Gamma(a+b)$ subject to $a$ and $b$ not being negative integers.

beta_e

These routines compute the Beta Function, $B(a,b) = \Gamma(a)\Gamma(b)/\Gamma(a+b)$ subject to $a$ and $b$ not being negative integers.

beta_inc

These routines compute the normalized incomplete Beta function $I_x(a,b) = B_x(a,b) / B(a,b)$ where

beta_inc_e

These routines compute the normalized incomplete Beta function $I_x(a,b) = B_x(a,b) / B(a,b)$ where

choose

These routines compute the combinatorial factor nchoosem$= n!/(m!(n-m)!)$

choose_e

These routines compute the combinatorial factor nchoosem$= n!/(m!(n-m)!)$

clausen

These routines compute the Clausen integral $Cl_2(x)$.

clausen_e

These routines compute the Clausen integral $Cl_2(x)$.

conicalP_0

These routines compute the conical function $P^0_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_0_e

These routines compute the conical function $P^0_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_1

These routines compute the conical function $P^1_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_1_e

These routines compute the conical function $P^1_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_cyl_reg

These routines compute the Regular Cylindrical Conical Function $P^{-m}_{-1/2 + i \lambda}(x)$ for $x > -1$ and $m \ge -1$.

conicalP_cyl_reg_e

These routines compute the Regular Cylindrical Conical Function $P^{-m}_{-1/2 + i \lambda}(x)$ for $x > -1$ and $m \ge -1$.

conicalP_half

These routines compute the irregular Spherical Conical Function $P^{1/2}_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_half_e

These routines compute the irregular Spherical Conical Function $P^{1/2}_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_mhalf

These routines compute the regular Spherical Conical Function $P^{-1/2}_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_mhalf_e

These routines compute the regular Spherical Conical Function $P^{-1/2}_{-1/2 + i \lambda}(x)$ for $x > -1$.

conicalP_sph_reg

These routines compute the Regular Spherical Conical Function $P^{-1/2-l}_{-1/2 + i \lambda}(x)$ for $x > -1$ and $l \ge -1$.

conicalP_sph_reg_e

These routines compute the Regular Spherical Conical Function $P^{-1/2-l}_{-1/2 + i \lambda}(x)$ for $x > -1$ and $l \ge -1$.

cos

These routines compute the cosine function $\cos(x)$.

cos_e

These routines compute the cosine function $\cos(x)$.

cos_err_e

This routine computes the cosine of an angle x with an associated absolute error dx, $\cos(x \pm dx)$. Note that this function is provided in the error-handling form only since its purpose is to compute the propagated error.

coulomb_CL_e

This function computes the Coulomb wave function normalization constant $C_L(\eta)$ for $L > -1$.

coupling_3j

These routines compute the Wigner 3-j coefficient,

coupling_3j_e

These routines compute the Wigner 3-j coefficient,

coupling_6j

These routines compute the Wigner 6-j coefficient,

coupling_6j_e

These routines compute the Wigner 6-j coefficient,

coupling_9j

These routines compute the Wigner 9-j coefficient,

coupling_9j_e

These routines compute the Wigner 9-j coefficient,

dawson

These routines compute the value of Dawson’s integral for x.

dawson_e

These routines compute the value of Dawson’s integral for x.

debye_1

These routines compute the first-order Debye function $D_1(x)$.

debye_1_e

These routines compute the first-order Debye function $D_1(x)$.

debye_2

These routines compute the second-order Debye function $D_2(x)$.

debye_2_e

These routines compute the second-order Debye function $D_2(x)$.

debye_3

These routines compute the third-order Debye function $D_3(x)$.

debye_3_e

These routines compute the third-order Debye function $D_3(x)$.

debye_4

These routines compute the fourth-order Debye function $D_4(x)$.

debye_4_e

These routines compute the fourth-order Debye function $D_4(x)$.

debye_5

These routines compute the fifth-order Debye function $D_5(x)$.

debye_5_e

These routines compute the fifth-order Debye function $D_5(x)$.

debye_6

These routines compute the sixth-order Debye function $D_6(x)$.

debye_6_e

These routines compute the sixth-order Debye function $D_6(x)$.

dilog

These routines compute the dilogarithm for a real argument. In Lewin’s notation this is $Li_2(x)$, the real part of the dilogarithm of a real $x$. It is defined by the integral representation

dilog_e

These routines compute the dilogarithm for a real argument. In Lewin’s notation this is $Li_2(x)$, the real part of the dilogarithm of a real $x$. It is defined by the integral representation

doublefact

These routines compute the double factorial $n!! = n(n-2)(n-4) \dots$. The maximum value of $n$ such that $n!!$ is not considered an overflow is given by the macro GSL_SF_DOUBLEFACT_NMAX and is 297.

doublefact_e

These routines compute the double factorial $n!! = n(n-2)(n-4) \dots$. The maximum value of $n$ such that $n!!$ is not considered an overflow is given by the macro GSL_SF_DOUBLEFACT_NMAX and is 297.

erf

These routines compute the error function $\text{erf}(x)$, where $\text{erf}(x) = (2/\sqrt{\pi}) \int_0^x dt \exp(-t^2)$.

erf_Q

These routines compute the upper tail of the Gaussian probability function $Q(x) = (1/\sqrt{2\pi}) \int_x^\infty dt \exp(-t^2/2)$

erf_Q_e

These routines compute the upper tail of the Gaussian probability function $Q(x) = (1/\sqrt{2\pi}) \int_x^\infty dt \exp(-t^2/2)$

erf_Z

These routines compute the Gaussian probability density function $Z(x) = (1/\sqrt{2\pi}) \exp(-x^2/2)$

erf_Z_e

These routines compute the Gaussian probability density function $Z(x) = (1/\sqrt{2\pi}) \exp(-x^2/2)$

erf_e

These routines compute the error function $\text{erf}(x)$, where $\text{erf}(x) = (2/\sqrt{\pi}) \int_0^x dt \exp(-t^2)$.

erfc

These routines compute the complementary error function $\text{erfc}(x) = 1 - \text{erf}(x) = (2/\sqrt{\pi}) \int_x^\infty \exp(-t^2)$

erfc_e

These routines compute the complementary error function $\text{erfc}(x) = 1 - \text{erf}(x) = (2/\sqrt{\pi}) \int_x^\infty \exp(-t^2)$

eta

These routines compute the eta function $\eta(s)$ for arbitrary s.

eta_e

These routines compute the eta function $\eta(s)$ for arbitrary s.

eta_int

These routines compute the eta function $\eta(n)$ for integer n.

eta_int_e

These routines compute the eta function $\eta(n)$ for integer n.

exp

These routines provide an exponential function $\exp(x)$ using GSL semantics and error checking.

exp_e

These routines provide an exponential function $\exp(x)$ using GSL semantics and error checking.

exp_err_e

This function exponentiates x with an associated absolute error dx.

exp_mult

These routines exponentiate x and multiply by the factor y to return the product $y \exp(x)$.

exp_mult_e

These routines exponentiate x and multiply by the factor y to return the product $y \exp(x)$.

exp_mult_err_e

This routine computes the product $y \exp(x)$ for the quantities x, y with associated absolute errors dx, dy.

expint_3

These routines compute the third-order exponential integral

expint_3_e

These routines compute the third-order exponential integral

expint_E1

These routines compute the exponential integral $E_1(x)$,

expint_E1_e

These routines compute the exponential integral $E_1(x)$,

expint_E2

These routines compute the second-order exponential integral $E_2(x)$,

expint_E2_e

These routines compute the second-order exponential integral $E_2(x)$,

expint_Ei

These routines compute the exponential integral $Ei(x)$,

expint_Ei_e

These routines compute the exponential integral $Ei(x)$,

expint_En

These routines compute the exponential integral $E_n(x)$ of order n,

expint_En_e

These routines compute the exponential integral $E_n(x)$ of order n,

expm1

These routines compute the quantity $\exp(x)-1$ using an algorithm that is accurate for small $x$.

expm1_e

These routines compute the quantity $\exp(x)-1$ using an algorithm that is accurate for small $x$.

exprel

These routines compute the quantity $(\exp(x)-1)/x$ using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion $(\exp(x)-1)/x = 1 + x/2 + x^2/(23) + x^3/(23*4) + \dots$.

exprel_2

These routines compute the quantity $2(\exp(x)-1-x)/x^2$ using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion $2(\exp(x)-1-x)/x^2 = 1 + x/3 + x^2/(34) + x^3/(34*5) + \dots$.

exprel_2_e

These routines compute the quantity $2(\exp(x)-1-x)/x^2$ using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion $2(\exp(x)-1-x)/x^2 = 1 + x/3 + x^2/(34) + x^3/(34*5) + \dots$.

exprel_e

These routines compute the quantity $(\exp(x)-1)/x$ using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion $(\exp(x)-1)/x = 1 + x/2 + x^2/(23) + x^3/(23*4) + \dots$.

exprel_n

These routines compute the $N$-relative exponential, which is the n-th generalization of the functions gsl_sf_exprel() and gsl_sf_exprel_2(). The $N$-relative exponential is given by,

exprel_n_e

These routines compute the $N$-relative exponential, which is the n-th generalization of the functions gsl_sf_exprel() and gsl_sf_exprel_2(). The $N$-relative exponential is given by,

fact

These routines compute the factorial $n!$. The factorial is related to the Gamma function by $n! = \Gamma(n+1)$. The maximum value of $n$ such that $n!$ is not considered an overflow is given by the macro GSL_SF_FACT_NMAX and is 170.

fact_e

These routines compute the factorial $n!$. The factorial is related to the Gamma function by $n! = \Gamma(n+1)$. The maximum value of $n$ such that $n!$ is not considered an overflow is given by the macro GSL_SF_FACT_NMAX and is 170.

fermi_dirac_0

These routines compute the complete Fermi-Dirac integral with an index of $0$. This integral is given by $F_0(x) = \ln(1 + e^x)$.

fermi_dirac_0_e

These routines compute the complete Fermi-Dirac integral with an index of $0$. This integral is given by $F_0(x) = \ln(1 + e^x)$.

fermi_dirac_1

These routines compute the complete Fermi-Dirac integral with an index of $1$, $F_1(x) = \int_0^\infty dt (t /(\exp(t-x)+1))$.

fermi_dirac_1_e

These routines compute the complete Fermi-Dirac integral with an index of $1$, $F_1(x) = \int_0^\infty dt (t /(\exp(t-x)+1))$.

fermi_dirac_2

These routines compute the complete Fermi-Dirac integral with an index of $2$, $F_2(x) = (1/2) \int_0^\infty dt (t^2 /(\exp(t-x)+1))$.

fermi_dirac_2_e

These routines compute the complete Fermi-Dirac integral with an index of $2$, $F_2(x) = (1/2) \int_0^\infty dt (t^2 /(\exp(t-x)+1))$.

fermi_dirac_3half

These routines compute the complete Fermi-Dirac integral $F_{3/2}(x)$.

fermi_dirac_3half_e

These routines compute the complete Fermi-Dirac integral $F_{3/2}(x)$.

fermi_dirac_half

These routines compute the complete Fermi-Dirac integral $F_{1/2}(x)$.

fermi_dirac_half_e

These routines compute the complete Fermi-Dirac integral $F_{1/2}(x)$.

fermi_dirac_inc_0

These routines compute the incomplete Fermi-Dirac integral with an index of zero, $F_0(x,b) = \ln(1 + e^{b-x}) - (b-x)$

fermi_dirac_inc_0_e

These routines compute the incomplete Fermi-Dirac integral with an index of zero, $F_0(x,b) = \ln(1 + e^{b-x}) - (b-x)$

fermi_dirac_int

These routines compute the complete Fermi-Dirac integral with an integer index of $j$, $F_j(x) = (1/\Gamma(j+1)) \int_0^\infty dt (t^j /(\exp(t-x)+1))$.

fermi_dirac_int_e

These routines compute the complete Fermi-Dirac integral with an integer index of $j$, $F_j(x) = (1/\Gamma(j+1)) \int_0^\infty dt (t^j /(\exp(t-x)+1))$.

fermi_dirac_m1

These routines compute the complete Fermi-Dirac integral with an index of $-1$. This integral is given by $F_{-1}(x) = e^x / (1 + e^x)$.

fermi_dirac_m1_e

These routines compute the complete Fermi-Dirac integral with an index of $-1$. This integral is given by $F_{-1}(x) = e^x / (1 + e^x)$.

fermi_dirac_mhalf

These routines compute the complete Fermi-Dirac integral $F_{-1/2}(x)$.

fermi_dirac_mhalf_e

These routines compute the complete Fermi-Dirac integral $F_{-1/2}(x)$.

gamma

These routines compute the Gamma function $\Gamma(x)$, subject to $x$ not being a negative integer or zero. The function is computed using the real Lanczos method. The maximum value of $x$ such that $\Gamma(x)$ is not considered an overflow is given by the macro GSL_SF_GAMMA_XMAX and is 171.0.

gamma_e

These routines compute the Gamma function $\Gamma(x)$, subject to $x$ not being a negative integer or zero. The function is computed using the real Lanczos method. The maximum value of $x$ such that $\Gamma(x)$ is not considered an overflow is given by the macro GSL_SF_GAMMA_XMAX and is 171.0.

gamma_inc

These functions compute the unnormalized incomplete Gamma Function $\Gamma(a,x) = \int_x^\infty dt t^{(a-1)} \exp(-t)$ for $a$ real and $x \ge 0$.

gamma_inc_P

These routines compute the complementary normalized incomplete Gamma Function $P(a,x) = 1 - Q(a,x) = 1/\Gamma(a) \int_0^x dt t^{(a-1)} \exp(-t)$ for $a > 0$, $x \ge 0$.

gamma_inc_P_e

These routines compute the complementary normalized incomplete Gamma Function $P(a,x) = 1 - Q(a,x) = 1/\Gamma(a) \int_0^x dt t^{(a-1)} \exp(-t)$ for $a > 0$, $x \ge 0$.

gamma_inc_Q

These routines compute the normalized incomplete Gamma Function $Q(a,x) = 1/\Gamma(a) \int_x^\infty dt t^{(a-1)} \exp(-t)$ for $a > 0$, $x \ge 0$.

gamma_inc_Q_e

These routines compute the normalized incomplete Gamma Function $Q(a,x) = 1/\Gamma(a) \int_x^\infty dt t^{(a-1)} \exp(-t)$ for $a > 0$, $x \ge 0$.

gamma_inc_e

These functions compute the unnormalized incomplete Gamma Function $\Gamma(a,x) = \int_x^\infty dt t^{(a-1)} \exp(-t)$ for $a$ real and $x \ge 0$.

gammainv

These routines compute the reciprocal of the gamma function, $1/\Gamma(x)$ using the real Lanczos method.

gammainv_e

These routines compute the reciprocal of the gamma function, $1/\Gamma(x)$ using the real Lanczos method.

gammastar

These routines compute the regulated Gamma Function $\Gamma^*(x)$ for $x > 0$. The regulated gamma function is given by,

gammastar_e

These routines compute the regulated Gamma Function $\Gamma^*(x)$ for $x > 0$. The regulated gamma function is given by,

gegenpoly_1

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_1_e

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_2

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_2_e

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_3

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_3_e

These functions evaluate the Gegenbauer polynomials $C^{(\lambda)}_n(x)$ using explicit representations for $n = 1, 2, 3$.

gegenpoly_n

These functions evaluate the Gegenbauer polynomial $C^{(\lambda)}_n(x)$ for a specific value of n, lambda, x subject to $\lambda > -1/2$, $n \ge 0$.

gegenpoly_n_e

These functions evaluate the Gegenbauer polynomial $C^{(\lambda)}_n(x)$ for a specific value of n, lambda, x subject to $\lambda > -1/2$, $n \ge 0$.

hazard

These routines compute the hazard function for the normal distribution.

hazard_e

These routines compute the hazard function for the normal distribution.

hermite

These routines evaluate the physicist Hermite polynomial $H_n(x)$ of order n at position x. If an overflow is detected, GSL_EOVRFLW is returned without calling the error handler.

hermite_deriv

These routines evaluate the m-th derivative of the physicist Hermite polynomial $H_n(x)$ of order n at position x.

hermite_deriv_e

These routines evaluate the m-th derivative of the physicist Hermite polynomial $H_n(x)$ of order n at position x.

hermite_e

These routines evaluate the physicist Hermite polynomial $H_n(x)$ of order n at position x. If an overflow is detected, GSL_EOVRFLW is returned without calling the error handler.

hermite_func

These routines evaluate the Hermite function $\psi_n(x)$ of order n at position x using a three term recurrence relation. The algorithm complexity is $O(n)$.

hermite_func_der

These routines evaluate the m-th derivative of the Hermite function $\psi_n(x)$ of order n at position x.

hermite_func_der_e

These routines evaluate the m-th derivative of the Hermite function $\psi_n(x)$ of order n at position x.

hermite_func_e

These routines evaluate the Hermite function $\psi_n(x)$ of order n at position x using a three term recurrence relation. The algorithm complexity is $O(n)$.

hermite_func_fast

These routines evaluate the Hermite function $\psi_n(x)$ of order n at position x using a the Cauchy integral algorithm due to Bunck, 2009. The algorithm complexity is $O(\sqrt{n})$.

hermite_func_fast_e

These routines evaluate the Hermite function $\psi_n(x)$ of order n at position x using a the Cauchy integral algorithm due to Bunck, 2009. The algorithm complexity is $O(\sqrt{n})$.

hermite_func_zero

These routines evaluate the s-th zero of the Hermite function $\psi_n(x)$ of order n.

hermite_func_zero_e

These routines evaluate the s-th zero of the Hermite function $\psi_n(x)$ of order n.

hermite_prob

These routines evaluate the probabilist Hermite polynomial $He_n(x)$ of order n at position x. If an overflow is detected, GSL_EOVRFLW is returned without calling the error handler.

hermite_prob_deriv

These routines evaluate the m-th derivative of the probabilist Hermite polynomial $He_n(x)$ of order n at position x.

hermite_prob_deriv_e

These routines evaluate the m-th derivative of the probabilist Hermite polynomial $He_n(x)$ of order n at position x.

hermite_prob_e

These routines evaluate the probabilist Hermite polynomial $He_n(x)$ of order n at position x. If an overflow is detected, GSL_EOVRFLW is returned without calling the error handler.

hermite_prob_zero

These routines evaluate the s-th zero of the probabilist Hermite polynomial $He_n(x)$ of order n.

hermite_prob_zero_e

These routines evaluate the s-th zero of the probabilist Hermite polynomial $He_n(x)$ of order n.

hermite_zero

These routines evaluate the s-th zero of the physicist Hermite polynomial $H_n(x)$ of order n.

hermite_zero_e

These routines evaluate the s-th zero of the physicist Hermite polynomial $H_n(x)$ of order n.

hydrogenicR

These routines compute the n-th normalized hydrogenic bound state radial wavefunction,

hydrogenicR_1

These routines compute the lowest-order normalized hydrogenic bound state radial wavefunction $R_1 := 2Z \sqrt{Z} \exp(-Z r)$.

hydrogenicR_1_e

These routines compute the lowest-order normalized hydrogenic bound state radial wavefunction $R_1 := 2Z \sqrt{Z} \exp(-Z r)$.

hydrogenicR_e

These routines compute the n-th normalized hydrogenic bound state radial wavefunction,

hyperg_0F1

These routines compute the hypergeometric function

hyperg_0F1_e

These routines compute the hypergeometric function

hyperg_1F1

These routines compute the confluent hypergeometric function

hyperg_1F1_e

These routines compute the confluent hypergeometric function

hyperg_1F1_int

These routines compute the confluent hypergeometric function

hyperg_1F1_int_e

These routines compute the confluent hypergeometric function

hyperg_2F0

These routines compute the hypergeometric function

hyperg_2F0_e

These routines compute the hypergeometric function

hyperg_2F1

These routines compute the Gauss hypergeometric function

hyperg_2F1_conj

These routines compute the Gauss hypergeometric function

hyperg_2F1_conj_e

These routines compute the Gauss hypergeometric function

hyperg_2F1_conj_renorm

These routines compute the renormalized Gauss hypergeometric function

hyperg_2F1_conj_renorm_e

These routines compute the renormalized Gauss hypergeometric function

hyperg_2F1_e

These routines compute the Gauss hypergeometric function

hyperg_2F1_renorm

These routines compute the renormalized Gauss hypergeometric function

hyperg_2F1_renorm_e

These routines compute the renormalized Gauss hypergeometric function

hyperg_U

These routines compute the confluent hypergeometric function $U(a,b,x)$.

hyperg_U_e

These routines compute the confluent hypergeometric function $U(a,b,x)$.

hyperg_U_int

These routines compute the confluent hypergeometric function $U(m,n,x)$ for integer parameters m, n.

hyperg_U_int_e

These routines compute the confluent hypergeometric function $U(m,n,x)$ for integer parameters m, n.

hypot

These routines compute the hypotenuse function $\sqrt{x^2 + y^2}$ avoiding overflow and underflow.

hypot_e

These routines compute the hypotenuse function $\sqrt{x^2 + y^2}$ avoiding overflow and underflow.

hzeta

These routines compute the Hurwitz zeta function $\zeta(s,q)$ for $s > 1$, $q > 0$.

hzeta_e

These routines compute the Hurwitz zeta function $\zeta(s,q)$ for $s > 1$, $q > 0$.

laguerre_1

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_1_e

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_2

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_2_e

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_3

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_3_e

These routines evaluate the generalized Laguerre polynomials $L^a_1(x)$, $L^a_2(x)$, $L^a_3(x)$ using explicit representations.

laguerre_n

These routines evaluate the generalized Laguerre polynomials $L^a_n(x)$ for $a > -1$, $n \ge 0$.

laguerre_n_e

These routines evaluate the generalized Laguerre polynomials $L^a_n(x)$ for $a > -1$, $n \ge 0$.

lambert_W0

These compute the principal branch of the Lambert W function, $W_0(x)$.

lambert_W0_e

These compute the principal branch of the Lambert W function, $W_0(x)$.

lambert_Wm1

These compute the secondary real-valued branch of the Lambert W function, $W_{-1}(x)$.

lambert_Wm1_e

These compute the secondary real-valued branch of the Lambert W function, $W_{-1}(x)$.

legendre_H3d

These routines compute the l-th radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space $\eta \ge 0$ and $l \ge 0$. In the flat limit this takes the form $L^{H3d}_l(\lambda,\eta) = j_l(\lambda\eta)$.

legendre_H3d_0

These routines compute the zeroth radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space,

legendre_H3d_0_e

These routines compute the zeroth radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space,

legendre_H3d_1

These routines compute the first radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space,

legendre_H3d_1_e

These routines compute the first radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space,

legendre_H3d_e

These routines compute the l-th radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space $\eta \ge 0$ and $l \ge 0$. In the flat limit this takes the form $L^{H3d}_l(\lambda,\eta) = j_l(\lambda\eta)$.

legendre_P1

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_P1_e

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_P2

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_P2_e

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_P3

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_P3_e

These functions evaluate the Legendre polynomials $P_l(x)$ using explicit representations for $l = 1, 2, 3$.

legendre_Pl

These functions evaluate the Legendre polynomial $P_l(x)$ for a specific value of l, x subject to $l \ge 0$ and $|x| \le 1$.

legendre_Pl_e

These functions evaluate the Legendre polynomial $P_l(x)$ for a specific value of l, x subject to $l \ge 0$ and $|x| \le 1$.

legendre_Plm

These routines compute the associated Legendre polynomial $P_l^m(x)$ for $m \ge 0$, $l \ge m$, and $|x| \le 1$.

legendre_Plm_e

These routines compute the associated Legendre polynomial $P_l^m(x)$ for $m \ge 0$, $l \ge m$, and $|x| \le 1$.

legendre_Q0

These routines compute the Legendre function $Q_0(x)$ for $x > -1$ and $x \ne 1$.

legendre_Q0_e

These routines compute the Legendre function $Q_0(x)$ for $x > -1$ and $x \ne 1$.

legendre_Q1

These routines compute the Legendre function $Q_1(x)$ for $x > -1$ and $x \ne 1$.

legendre_Q1_e

These routines compute the Legendre function $Q_1(x)$ for $x > -1$ and $x \ne 1$.

legendre_Ql

These routines compute the Legendre function $Q_l(x)$ for $x > -1$, $x \ne 1$ and $l \ge 0$.

legendre_Ql_e

These routines compute the Legendre function $Q_l(x)$ for $x > -1$, $x \ne 1$ and $l \ge 0$.

legendre_array_size

This function is now deprecated and will be removed in a future release.

legendre_sphPlm

These routines compute the normalized associated Legendre polynomial $\sqrt{(2l+1)/(4\pi)} \sqrt{(l-m)!/(l+m)!} P_l^m(x)$ suitable for use in spherical harmonics. The parameters must satisfy $m \ge 0$, $l \ge m$, and $|x| \le 1$. These routines avoid the overflows that occur for the standard normalization of $P_l^m(x)$.

legendre_sphPlm_e

These routines compute the normalized associated Legendre polynomial $\sqrt{(2l+1)/(4\pi)} \sqrt{(l-m)!/(l+m)!} P_l^m(x)$ suitable for use in spherical harmonics. The parameters must satisfy $m \ge 0$, $l \ge m$, and $|x| \le 1$. These routines avoid the overflows that occur for the standard normalization of $P_l^m(x)$.

ln_beta

These routines compute the logarithm of the Beta Function, $\log(B(a,b))$ subject to $a$ and $b$ not being negative integers.

ln_beta_e

These routines compute the logarithm of the Beta Function, $\log(B(a,b))$ subject to $a$ and $b$ not being negative integers.

ln_choose

These routines compute the logarithm of nchoosem. This is equivalent to the sum $\log(n!) - \log(m!) - \log((n-m)!)$.

ln_choose_e

These routines compute the logarithm of nchoosem. This is equivalent to the sum $\log(n!) - \log(m!) - \log((n-m)!)$.

ln_cosh

These routines compute $\log(\cosh(x))$ for any x.

ln_cosh_e

These routines compute $\log(\cosh(x))$ for any x.

ln_doublefact

These routines compute the logarithm of the double factorial of n, $\log(n!!)$.

ln_doublefact_e

These routines compute the logarithm of the double factorial of n, $\log(n!!)$.

ln_fact

These routines compute the logarithm of the factorial of n, $\log(n!)$. The algorithm is faster than computing $\ln(\Gamma(n+1))$ via gsl_sf_lngamma() for $n < 170$, but defers for larger n.

ln_fact_e

These routines compute the logarithm of the factorial of n, $\log(n!)$. The algorithm is faster than computing $\ln(\Gamma(n+1))$ via gsl_sf_lngamma() for $n < 170$, but defers for larger n.

ln_gamma

These routines compute the logarithm of the Gamma function, $\log(\Gamma(x))$, subject to $x$ not being a negative integer or zero. For $x < 0$ the real part of $\log(\Gamma(x))$ is returned, which is equivalent to $\log(|\Gamma(x)|)$. The function is computed using the real Lanczos method.

ln_gamma_e

These routines compute the logarithm of the Gamma function, $\log(\Gamma(x))$, subject to $x$ not being a negative integer or zero. For $x < 0$ the real part of $\log(\Gamma(x))$ is returned, which is equivalent to $\log(|\Gamma(x)|)$. The function is computed using the real Lanczos method.

ln_poch

These routines compute the logarithm of the Pochhammer symbol, $\log((a)_x) = \log(\Gamma(a + x)/\Gamma(a))$.

ln_poch_e

These routines compute the logarithm of the Pochhammer symbol, $\log((a)_x) = \log(\Gamma(a + x)/\Gamma(a))$.

ln_sinh

These routines compute $\log(\sinh(x))$ for $x > 0$.

ln_sinh_e

These routines compute $\log(\sinh(x))$ for $x > 0$.

log

These routines compute the logarithm of x, $\log(x)$, for $x > 0$.

log_1plusx

These routines compute $\log(1 + x)$ for $x > -1$ using an algorithm that is accurate for small x.

log_1plusx_e

These routines compute $\log(1 + x)$ for $x > -1$ using an algorithm that is accurate for small x.

log_1plusx_mx

These routines compute $\log(1 + x) - x$ for $x > -1$ using an algorithm that is accurate for small x.

log_1plusx_mx_e

These routines compute $\log(1 + x) - x$ for $x > -1$ using an algorithm that is accurate for small x.

log_abs

These routines compute the logarithm of the magnitude of x, $\log(|x|)$, for $x \ne 0$.

log_abs_e

These routines compute the logarithm of the magnitude of x, $\log(|x|)$, for $x \ne 0$.

log_e

These routines compute the logarithm of x, $\log(x)$, for $x > 0$.

log_erfc

These routines compute the logarithm of the complementary error function $\log(\text{erfc}(x))$.

log_erfc_e

These routines compute the logarithm of the complementary error function $\log(\text{erfc}(x))$.

mathieu_Mc

These routines compute the radial j-th kind Mathieu functions $Mc_n^{(j)}(q,x)$ and $Ms_n^{(j)}(q,x)$ of order n.

mathieu_Mc_e

These routines compute the radial j-th kind Mathieu functions $Mc_n^{(j)}(q,x)$ and $Ms_n^{(j)}(q,x)$ of order n.

mathieu_Ms

These routines compute the radial j-th kind Mathieu functions $Mc_n^{(j)}(q,x)$ and $Ms_n^{(j)}(q,x)$ of order n.

mathieu_Ms_e

These routines compute the radial j-th kind Mathieu functions $Mc_n^{(j)}(q,x)$ and $Ms_n^{(j)}(q,x)$ of order n.

mathieu_a

These routines compute the characteristic values $a_n(q)$, $b_n(q)$ of the Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_a_e

These routines compute the characteristic values $a_n(q)$, $b_n(q)$ of the Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_b

These routines compute the characteristic values $a_n(q)$, $b_n(q)$ of the Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_b_e

These routines compute the characteristic values $a_n(q)$, $b_n(q)$ of the Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_ce

These routines compute the angular Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_ce_e

These routines compute the angular Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_se

These routines compute the angular Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

mathieu_se_e

These routines compute the angular Mathieu functions $ce_n(q,x)$ and $se_n(q,x)$, respectively.

multiply

This function multiplies x and y storing the product and its associated error in result.

multiply_e

This function multiplies x and y storing the product and its associated error in result.

multiply_err_e

This function multiplies x and y with associated absolute errors dx and dy. The product $xy \pm xy \sqrt{(dx/x)^2 +(dy/y)^2}$ is stored in result.

poch

These routines compute the Pochhammer symbol $(a)_x = \Gamma(a + x)/\Gamma(a)$. The Pochhammer symbol is also known as the Apell symbol and sometimes written as $(a,x)$. When $a$ and $a + x$ are negative integers or zero, the limiting value of the ratio is returned.

poch_e

These routines compute the Pochhammer symbol $(a)_x = \Gamma(a + x)/\Gamma(a)$. The Pochhammer symbol is also known as the Apell symbol and sometimes written as $(a,x)$. When $a$ and $a + x$ are negative integers or zero, the limiting value of the ratio is returned.

pochrel

These routines compute the relative Pochhammer symbol $((a)_x - 1)/x$ where $(a)_x = \Gamma(a + x)/\Gamma(a)$.

pochrel_e

These routines compute the relative Pochhammer symbol $((a)_x - 1)/x$ where $(a)_x = \Gamma(a + x)/\Gamma(a)$.

pow_int

These routines compute the power $x^n$ for integer n. The power is computed using the minimum number of multiplications. For example, $x^8$ is computed as $((x^2)^2)^2$, requiring only 3 multiplications. For reasons of efficiency, these functions do not check for overflow or underflow conditions. The following is a simple example:

pow_int_e

These routines compute the power $x^n$ for integer n. The power is computed using the minimum number of multiplications. For example, $x^8$ is computed as $((x^2)^2)^2$, requiring only 3 multiplications. For reasons of efficiency, these functions do not check for overflow or underflow conditions. The following is a simple example:

psi

These routines compute the digamma function $\psi(x)$ for general x, $x \ne 0$.

psi_1

These routines compute the Trigamma function $\psi’(x)$ for general x.

psi_1_e

These routines compute the Trigamma function $\psi’(x)$ for general x.

psi_1_int

These routines compute the Trigamma function $\psi’(n)$ for positive integer $n$.

psi_1_int_e

These routines compute the Trigamma function $\psi’(n)$ for positive integer $n$.

psi_1piy

These routines compute the real part of the digamma function on the line $1 + i y$, $\Re[\psi(1 + i y)]$.

psi_1piy_e

These routines compute the real part of the digamma function on the line $1 + i y$, $\Re[\psi(1 + i y)]$.

psi_e

These routines compute the digamma function $\psi(x)$ for general x, $x \ne 0$.

psi_int

These routines compute the digamma function $\psi(n)$ for positive integer n. The digamma function is also called the Psi function.

psi_int_e

These routines compute the digamma function $\psi(n)$ for positive integer n. The digamma function is also called the Psi function.

psi_n

These routines compute the polygamma function $\psi^{(n)}(x)$ for $n \ge 0$, $x > 0$.

psi_n_e

These routines compute the polygamma function $\psi^{(n)}(x)$ for $n \ge 0$, $x > 0$.

sin

These routines compute the sine function $\sin(x)$.

sin_e

These routines compute the sine function $\sin(x)$.

sin_err_e

This routine computes the sine of an angle x with an associated absolute error dx, $\sin(x \pm dx)$. Note that this function is provided in the error-handling form only since its purpose is to compute the propagated error.

sinc

These routines compute $\text{sinc}(x) = \sin(\pi x) / (\pi x)$ for any value of x.

sinc_e

These routines compute $\text{sinc}(x) = \sin(\pi x) / (\pi x)$ for any value of x.

synchrotron_1

These routines compute the first synchrotron function $x \int_x^\infty dt K_{5/3}(t)$ for $x \ge 0$.

synchrotron_1_e

These routines compute the first synchrotron function $x \int_x^\infty dt K_{5/3}(t)$ for $x \ge 0$.

synchrotron_2

These routines compute the second synchrotron function $x K_{2/3}(x)$ for $x \ge 0$.

synchrotron_2_e

These routines compute the second synchrotron function $x K_{2/3}(x)$ for $x \ge 0$.

taylorcoeff

These routines compute the Taylor coefficient $x^n / n!$ for $x \ge 0$, $n \ge 0$

taylorcoeff_e

These routines compute the Taylor coefficient $x^n / n!$ for $x \ge 0$, $n \ge 0$

transport_2

These routines compute the transport function $J(2,x)$.

transport_2_e

These routines compute the transport function $J(2,x)$.

transport_3

These routines compute the transport function $J(3,x)$.

transport_3_e

These routines compute the transport function $J(3,x)$.

transport_4

These routines compute the transport function $J(4,x)$.

transport_4_e

These routines compute the transport function $J(4,x)$.

transport_5

These routines compute the transport function $J(5,x)$.

transport_5_e

These routines compute the transport function $J(5,x)$.

zeta

These routines compute the Riemann zeta function $\zeta(s)$ for arbitrary s, $s \ne 1$.

zeta_e

These routines compute the Riemann zeta function $\zeta(s)$ for arbitrary s, $s \ne 1$.

zeta_int

These routines compute the Riemann zeta function $\zeta(n)$ for integer n, $n \ne 1$.

zeta_int_e

These routines compute the Riemann zeta function $\zeta(n)$ for integer n, $n \ne 1$.

zeta_minus_1

These routines compute $\zeta(s) - 1$ for arbitrary s, $s \ne 1$.

zeta_minus_1_e

These routines compute $\zeta(s) - 1$ for arbitrary s, $s \ne 1$.

zeta_minus_1_int

These routines compute $\zeta(n) - 1$ for integer n, $n \ne 1$.

zeta_minus_1_int_e

These routines compute $\zeta(n) - 1$ for integer n, $n \ne 1$.